modularisation of the site, has seen alot of code moved from main.js to navbar.js. This is frustrating as navbar looks more and more like main.js was. Consider what can be moved back to main and why I moved it? Behaviour is coupled strongly this needs un-coupled. Currently I have moved some thumbnail behaviour into navbar.js this should be the case. ?have a look at exact behaviourand it's function code and what else is called, can i separate navbar behaviour from thumbnail behaviour?
