+++ index.html
@@ -15,6 +15,7 @@
   <script src="https://code.jquery.com/jquery-1.11.0.min.js" defer></script>
   
   <!-- <script src="js/mainNoModul.js" defer></script> -->
+  <script type="module" src="js/navbar.js"></script>
   <script type="module" defer src="js/mainModular.js" ></script>
   <script nomodule defer src="js/mainNonModular.js" ></script>
 
@@ -97,7 +98,7 @@
     </nav>
 
     <div class="svg-container container-fluid" id="mountains">
-      <svg viewBox="0 0 1200 800" id="svg" xmlns="https://www.w3.org/2000/svg">
+      <svg viewBox="0 0 1200 800" id="svg" xmlns="https://www.w3.org/2000/svg" visibility = "hidden">
         <mask id="m">
           <g class="img-fluid" id="cloud1">
             <rect fill="#fff" width="1200" height="800" y="1200" />
