+++ js/navbar.js
@@ -3,16 +3,18 @@
 // Handles the sticky navbar logic
 
 // Main function to initialize the navbar
-
+import { getDomElements } from "./domUtils.js";
+const domElements = getDomElements(); 
 let endTopY = 0;
-export function setupNavbar(domElements, triggerOffset = 320) {
+
+const totalThumbWidth = getThumbWidthWithMargin();
+
+
+export function setupNavbar( domElements, triggerOffset = 320) {
   if (!domElements) {
     console.error("❌ Navbar elements missing!");
     return;
   }
-
-  console.log("✅ Initializing Navbar with elements:", domElements);
-
   enableStickyNavbar(triggerOffset); // Make navbar sticky
   setupNavbarEvents(domElements); // Attach event listeners
   setupDynamicLinks(); // Assign dynamic links
@@ -44,10 +46,8 @@ export function enableStickyNavbar(triggerOffset) {
       (isLandscapeSmall && isSmallHeight && inSmallHeightScrollRange)
     ) {
       $(".navbar").addClass("sticky");
-      console.log("Navbar is now sticky"); // ✅ Debugging
     } else {
       $(".navbar").removeClass("sticky");
-      console.log("Navbar is NOT sticky"); // ✅ Debugging
     }
   });
 }
@@ -70,6 +70,7 @@ export function setupNavbarEvents(domElements) {
     return;
   }
 
+  console.log("✅ domElements before setupDropdownHover:", domElements);
   // Setup dropdown hover behaviour
   setupDropdownHover(domElements.navSoftware, domElements.navDropMenuSoftware, domElements.thumbSoft);
   setupDropdownHover(domElements.navMotion, domElements.navDropMenuMotion, domElements.thumbMot);
@@ -89,24 +90,22 @@ export function setupNavbarEvents(domElements) {
 }
 
 // 🔹 Handles dropdown hover behavior
-function setupDropdownHover(navItem, dropdownMenu, thumbElement) {
+function setupDropdownHover(navItem, dropdownMenu, thumbElements) {
   if (!navItem || !dropdownMenu) {
-    console.error("❌ Missing required navbar elements:", { navItem, dropdownMenu, thumbElement });
+    console.error("❌ Missing required navbar elements:", { navItem, dropdownMenu, thumbElements });
     return;
   }
-
-  console.log(`✅ Setting up dropdown hover for: ${navItem.id}`);
 
   navItem.addEventListener("mouseenter", function () {
     showDropMenu(dropdownMenu);
     navItem.classList.add("active");
-    if (thumbElement) thumbElement.classList.add("active"); // ✅ Only add if exists
+    if (thumbElements) thumbElements.classList.add("active"); // ✅ Only add if exists
   });
 
   navItem.addEventListener("mouseleave", function () {
     delayedHide(dropdownMenu);
     navItem.classList.remove("active");
-    if (thumbElement) thumbElement.classList.remove("active"); // ✅ Only remove if exists
+    if (thumbElements) thumbElements.classList.remove("active"); // ✅ Only remove if exists
   });
 
   dropdownMenu.addEventListener("mouseenter", cancelHide);
@@ -166,12 +165,24 @@ export function autoScrollNow() {
 }
 
 // Displays WIP message
-export function showWip(thumbElement) {
-  updateWIPDimensions(getEndTopY(), thumbElement);
+export function showWip() {
+  // const domElements = getDomElements(); // ✅ Ensure we have latest elements
+
+  console.log("✅ domElements in showWip:", domElements); // ✅ Debugging
+
+  if (!domElements.thumbElements || domElements.thumbElements.length === 0) {
+    console.error("❌ Error: thumbElements is missing or empty!", domElements.thumbElements);
+    return;
+  }
+  updateWIPDimensions(getEndTopY(), domElements.thumbElements);
   updateDimensionsNoMargins();
   document.getElementById("wip").style.display = "block";
-  document.getElementById("contactForm").style.display = "none";
 }
+//   updateWIPDimensions(getEndTopY(), thumbElements);
+//   updateDimensionsNoMargins();
+//   document.getElementById("wip").style.display = "block";
+//   document.getElementById("contactForm").style.display = "none";
+// }
 
 
 // Hides WIP message
@@ -189,31 +200,123 @@ export function showScrollBar() {
   document.documentElement.style.overflow = ""; // Show scroll on the entire document
 }
 
-export function updateWIPDimensions(endTopY) {
+export function updateWIPDimensions(endTopY, thumbElements) {
   const wip = document.querySelector(".wip .box");
 
-  if (!thumbElement) {
-    console.error("❌ thumbElement is undefined in updateWIPDimensions!");
-  }
-  if (!wip) {
-    console.error("❌ wip is undefined in updateWIPDimensions!");
-  }
-
-
-  if (!wip || !thumbElement){
-    console.error("Missing elements:", { wip, thumbElement });
+  if (!thumbElements || thumbElements.length === 0) {
+    console.error("❌ Error: thumbElements is missing or empty! in updateWIPDimensions()-navbar.js", thumbElements);
     return;
   }
+  if (!wip) {
+    console.error("Error: wip is undefined in updateWIPDimensions!");
+    return
+  }
+  if (!wip || !thumbElements){
+    console.error("Missing elements:", { wip, thumbElements });
+    return;
+  }
+
+  // const thumbWidthWithoutMargin = getThumbWidthWithoutMargin();
+
+  // // Calculate new width and height for the modal box
+  // const newWidth = Math.max(thumbWidthWithoutMargin * 2 + 4, 300);
+  // const newHeight = newWidth; // Assuming we want a square wip
+
+  // // Update modal dimensions
+  // wip.style.width = `${newWidth}px`;
+  // wip.style.height = `${newHeight}px`;
+
+  // // Calculate the center of the screen
+  // const centerX = window.innerWidth / 2 + 6; //refining positioning
+
+  // // Calculate the new left position to center the modal box
+  // const newLeft = centerX - newWidth / 2;
+
+  // // Use the passed endTopY for the new top position
+  // const newTop = endTopY + 12.5; //  works for alignment on y axis
+
+  // // Update modal position
+  // wip.style.position = "absolute";
+  // wip.style.left = `${newLeft}px`;
+  // wip.style.top = `${newTop}px`;
+
+  // // modalBox.style.display = "block";
+
+  const thumbWidthWithoutMargin = getThumbWidthWithoutMargin();
+  const newWidth = Math.max(thumbWidthWithoutMargin * 2 + 4, 300);
+  const newHeight = newWidth; // Assuming a square WIP
+
+  wip.style.width = `${newWidth}px`;
+  wip.style.height = `${newHeight}px`;
+
+  const centerX = window.innerWidth / 2 + 6;
+  const newLeft = centerX - newWidth / 2;
+  const newTop = endTopY + 12.5;
+
+  wip.style.position = "absolute";
+  wip.style.left = `${newLeft}px`;
+  wip.style.top = `${newTop}px`;
+
+  console.log("✅ WIP dimensions updated with thumbElements.");
+}
+
+export function getThumbWidthWithoutMargin() {
+  // const domElements = getDomElements();
+  const computedStyle = window.getComputedStyle(domElements.thumbElements[0]);
+  const thumbWidth = parseFloat(computedStyle.getPropertyValue("width"));
+
+  return thumbWidth;
+}
+
+export function getThumbMargin() {
+  const computedStyle = window.getComputedStyle(domElements.thumbElements[0]);
+  const thumbMargin = parseFloat(
+    computedStyle.getPropertyValue("margin-right")
+  );
+  return thumbMargin;
+}
+
+export function getThumbWidthWithMargin() {
+  const computedStyle = window.getComputedStyle(domElements.thumbElements[0]);
+  const thumbWidth = parseFloat(computedStyle.getPropertyValue("width"));
+  const thumbMargin = parseFloat(
+    computedStyle.getPropertyValue("margin-right")
+  );
+
+  return thumbWidth + thumbMargin * 2;
+}
+
+function getComputedStyleValue(element, property) {
+
+  return parseInt(window.getComputedStyle(element).getPropertyValue(property));
+}
+
+let screenWidthHalved = svg.viewBox.baseVal.width / 2;
+let screenHeightHalved = svg.viewBox.baseVal.height / 2;
+export let endLeftX, endRightX, endBottomY; 
+let thumbWidth = getComputedStyleValue(
+  document.querySelector(".thumbShape"),
+  "width"
+);
+
+export const landscapeMediaQuery = window.matchMedia(
+  "(orientation: landscape) and (max-width: 991.98px) and (max-height: 600px)"
+);
+
+export function updateModalDimensions(endTopY) {
+  const modalBox = document.querySelector(".modalbox .box");
+
+  if (!modalBox || !domElements.thumbElements[0]) return;
 
   const thumbWidthWithoutMargin = getThumbWidthWithoutMargin();
 
   // Calculate new width and height for the modal box
   const newWidth = Math.max(thumbWidthWithoutMargin * 2 + 4, 300);
-  const newHeight = newWidth; // Assuming we want a square wip
+  const newHeight = newWidth; // Assuming we want a square modal
 
   // Update modal dimensions
-  wip.style.width = `${newWidth}px`;
-  wip.style.height = `${newHeight}px`;
+  modalBox.style.width = `${newWidth}px`;
+  modalBox.style.height = `${newHeight}px`;
 
   // Calculate the center of the screen
   const centerX = window.innerWidth / 2 + 6; //refining positioning
@@ -222,16 +325,158 @@ export function updateWIPDimensions(endTopY) {
   const newLeft = centerX - newWidth / 2;
 
   // Use the passed endTopY for the new top position
-  const newTop = endTopY + 12.5; //  works for alignment on y axis
+  const newTop = getEndTopY() + 12.5; //  works for alignment on y axis
 
   // Update modal position
-  wip.style.position = "absolute";
-  wip.style.left = `${newLeft}px`;
-  wip.style.top = `${newTop}px`;
+  modalBox.style.position = "absolute";
+  modalBox.style.left = `${newLeft}px`;
+  modalBox.style.top = `${newTop}px`;
 
   // modalBox.style.display = "block";
 }
 
+
+
+// Contact form sizing and positioning
+export function formControl(endTopY) {
+  const contactForm = document.querySelector(".formDiv#contactForm");
+
+  if (!contactForm) return;
+
+  const computedStyleForm = window.getComputedStyle(contactForm);
+  const formWidth = parseFloat(computedStyleForm.getPropertyValue("width"));
+  const formHeight = parseFloat(computedStyleForm.getPropertyValue("height"));
+
+  // Calculate form position
+  const formX = window.innerWidth / 2 - formWidth / 2 + 6.75;
+  const formY = getEndTopY() + 12; //+ 190;
+
+  // Update modal position
+  contactForm.style.position = "absolute";
+  contactForm.style.left = `${formX}px`;
+  contactForm.style.top = `${formY}px`;
+}
+
+
+export function updateDimensions() {
+   thumbWidth = Math.min(300, window.innerWidth / 6);
+  screenWidthHalved = window.innerWidth / 2;
+  screenHeightHalved = window.innerHeight / 2;
+
+  endLeftX = screenWidthHalved - totalThumbWidth;
+
+  updateEndTopY(); // ✅ Update the value first
+  const updatedEndTopY = getEndTopY(); // ✅ Retrieve the updated value
+
+  if (updatedEndTopY === undefined) {
+    console.error("Error: getEndTopY() returned undefined!");
+    return; // Stop execution if the value is not set
+  }
+
+  const endTopY = updatedEndTopY; // ✅ Use a local variable, do not redeclare globally
+
+  endRightX = screenWidthHalved;
+  updateEndBottomY();
+}
+
+export function spaceoutThumbs() {
+  gsap.to("#software", {
+    x: endLeftX,
+    y: getEndTopY(),
+    duration: 1,
+    ease: "power2.out",
+  });
+  gsap.to("#photography", {
+    x: endRightX,
+    y: getEndTopY(),
+    duration: 1,
+    ease: "power2.out",
+  });
+  gsap.to("#diy", {
+    x: endRightX,
+    y: endBottomY,
+    duration: 1,
+    ease: "power2.out",
+  });
+  gsap.to("#motion", {
+    x: endLeftX,
+    y: endBottomY,
+    duration: 1,
+    ease: "power2.out",
+  });
+}
+
+
+export function updateDimensionsNoMargins() {
+  setTimeout(() => {
+    const thumbWidth = Math.min(300, window.innerWidth / 6);
+    console.log("thumbWidth: ", thumbWidth);
+    const screenWidthHalved = window.innerWidth / 2;
+    const screenHeightHalved = window.innerHeight / 2;
+    const widthThumb = getThumbWidthWithoutMargin();
+    const marginWidth = getThumbMargin();
+
+    updateEndTopY();
+
+    const updatedEndTopY = getEndTopY(); // ✅ Fetch dynamically
+    if (updatedEndTopY === undefined) {
+      console.error("Error: getEndTopY() returned undefined!");
+      return; 
+    }
+
+    const endTopY = updatedEndTopY + marginWidth - 15; 
+    const endLeftX = screenWidthHalved - (widthThumb + marginWidth);
+    const endRightX = screenWidthHalved - marginWidth;
+    const endBottomY = updateEndBottomY() - marginWidth - 15;
+
+    collectThumbs();
+    updateModalDimensions(endTopY);
+    formControl(endTopY);
+  }, 450); // delay to ensure scrollbar removal takes effect
+}
+
+export function updateEndBottomY() {
+  if (
+    window.matchMedia("(orientation: landscape) and (max-width: 991.98px)")
+      .matches
+  ) {
+    endBottomY = window.innerHeight * 1.275 + totalThumbWidth; // Adjust multiplier for this condition
+  } else {
+    endBottomY = window.innerHeight * 1.325 + totalThumbWidth;
+  }
+  return endBottomY;
+}
+
+export function collectThumbs() {
+  gsap.to("#software", {
+    scale: 1,
+    x: endLeftX,
+    y: getEndTopY(),
+    duration: 1,
+    ease: "power2.out",
+  });
+  gsap.to("#photography", {
+    scale: 1,
+    x: endRightX,
+    y: getEndTopY(),
+    duration: 1,
+    ease: "power2.out",
+  });
+  gsap.to("#diy", {
+    scale: 1,
+    x: endRightX,
+    y: endBottomY,
+    duration: 1,
+    ease: "power2.out",
+  });
+  gsap.to("#motion", {
+    scale: 1,
+    x: endLeftX,
+    y: endBottomY,
+    duration: 1,
+    ease: "power2.out",
+  });
+}
 
 
 
@@ -241,12 +486,10 @@ export function updateEndTopY() {
   } else {
     endTopY = window.innerHeight * 1.325;
   }
-  console.log("Updated endTopY:", endTopY); // Debugging
   return endTopY;
 }
 
 export function getEndTopY() {
-  console.log("Fetching endTopY:", endTopY); // Debugging
   return endTopY;
 }
